Creates HtmlStringUtilities for the string helper methods and wraps the logic in UmbracoHelper
