+++ src/Umbraco.Web/UmbracoHelper.cs
@@ -42,6 +42,7 @@ namespace Umbraco.Web
         private readonly IUmbracoComponentRenderer _componentRenderer;
         private PublishedContentQuery _query;
         private readonly MembershipHelper _membershipHelper;
+        private readonly HtmlStringUtilities _stringUtilities = new HtmlStringUtilities();
         private ITagQuery _tag;
         private readonly IDataTypeService _dataTypeService;
         private readonly UrlProvider _urlProvider;
@@ -885,10 +886,7 @@ namespace Umbraco.Web
 		/// <returns>The text with text line breaks replaced with html linebreaks (<br/>)</returns>
 		public string ReplaceLineBreaksForHtml(string text)
 		{
-			if (bool.Parse(Umbraco.Core.Configuration.GlobalSettings.EditXhtmlMode))
-				return text.Replace("\n", "<br/>\n");
-			else
-				return text.Replace("\n", "<br />\n");
+            return _stringUtilities.ReplaceLineBreaksForHtml(text);
 		}
 
 		/// <summary>
@@ -911,79 +909,22 @@ namespace Umbraco.Web
 		}
 		public HtmlString StripHtml(string html, params string[] tags)
 		{
-			return StripHtmlTags(html, tags);
+            return _stringUtilities.StripHtmlTags(html, tags);
 		}
-
-		private HtmlString StripHtmlTags(string html, params string[] tags)
-		{
-			var doc = new HtmlDocument();
-			doc.LoadHtml("<p>" + html + "</p>");
-            var targets = new List<HtmlNode>();
-
-            var nodes = doc.DocumentNode.FirstChild.SelectNodes(".//*");
-            if (nodes != null)
-            {
-                foreach (var node in nodes)
-                {
-                    //is element
-                    if (node.NodeType != HtmlNodeType.Element) continue;
-                    var filterAllTags = (tags == null || !tags.Any());
-                    if (filterAllTags || tags.Any(tag => string.Equals(tag, node.Name, StringComparison.CurrentCultureIgnoreCase)))
-                    {
-                        targets.Add(node);
-                    }
-                }
-                foreach (var target in targets)
-                {
-                    HtmlNode content = doc.CreateTextNode(target.InnerText);
-                    target.ParentNode.ReplaceChild(content, target);
-                }
-            }
-            else
-            {
-                return new HtmlString(html);
-            }
-            return new HtmlString(doc.DocumentNode.FirstChild.InnerHtml);
-		}
-
+        
 		public string Coalesce(params object[] args)
 		{
-			return Coalesce<DynamicNull>(args);
-		}
-
-		internal string Coalesce<TIgnore>(params object[] args)
-		{
-			foreach (var sArg in args.Where(arg => arg != null && arg.GetType() != typeof(TIgnore)).Select(arg => string.Format("{0}", arg)).Where(sArg => !string.IsNullOrWhiteSpace(sArg)))
-			{
-				return sArg;
-			}
-			return string.Empty;
+            return _stringUtilities.Coalesce<DynamicNull>(args);
 		}
 
 		public string Concatenate(params object[] args)
 		{
-			return Concatenate<DynamicNull>(args);
-		}
-
-		internal string Concatenate<TIgnore>(params object[] args)
-		{
-			var result = new StringBuilder();
-			foreach (var sArg in args.Where(arg => arg != null && arg.GetType() != typeof(TIgnore)).Select(arg => string.Format("{0}", arg)).Where(sArg => !string.IsNullOrWhiteSpace(sArg)))
-			{
-				result.Append(sArg);
-			}
-			return result.ToString();
+            return _stringUtilities.Concatenate<DynamicNull>(args);
 		}
 
 		public string Join(string seperator, params object[] args)
 		{
-			return Join<DynamicNull>(seperator, args);
-		}
-
-		internal string Join<TIgnore>(string seperator, params object[] args)
-		{
-			var results = args.Where(arg => arg != null && arg.GetType() != typeof(TIgnore)).Select(arg => string.Format("{0}", arg)).Where(sArg => !string.IsNullOrWhiteSpace(sArg)).ToList();
-			return string.Join(seperator, results);
+            return _stringUtilities.Join<DynamicNull>(seperator, args);
 		}
 
 		public IHtmlString Truncate(IHtmlString html, int length)
@@ -1020,153 +961,7 @@ namespace Umbraco.Web
 		}
 		public IHtmlString Truncate(string html, int length, bool addElipsis, bool treatTagsAsContent)
 		{
-            using (var outputms = new MemoryStream())
-            {
-                using (var outputtw = new StreamWriter(outputms))
-                {
-                    using (var ms = new MemoryStream())
-                    {
-                        using (var tw = new StreamWriter(ms))
-                        {
-                            tw.Write(html);
-                            tw.Flush();
-                            ms.Position = 0;
-                            var tagStack = new Stack<string>();
-
-                            using (TextReader tr = new StreamReader(ms))
-                            {
-                                bool isInsideElement = false,
-                                    lengthReached = false,
-                                    insideTagSpaceEncountered = false,
-                                    isTagClose = false;
-
-                                int ic = 0,
-                                    currentLength = 0,
-                                    currentTextLength = 0;
-
-                                string currentTag = string.Empty,
-                                    tagContents = string.Empty;
-
-                                while ((ic = tr.Read()) != -1)
-                                {
-                                    bool write = true;
-
-                                    switch ((char)ic)
-                                    {
-                                        case '<':
-                                            if (!lengthReached)
-                                            {
-                                                isInsideElement = true;
-                                            }
-
-                                            insideTagSpaceEncountered = false;
-                                            currentTag = string.Empty;
-                                            tagContents = string.Empty;
-                                            isTagClose = false;
-                                            if (tr.Peek() == (int)'/')
-                                            {
-                                                isTagClose = true;
-                                            }
-                                            break;
-
-                                        case '>':
-                                            isInsideElement = false;
-
-                                            if (isTagClose && tagStack.Count > 0)
-                                            {
-                                                string thisTag = tagStack.Pop();
-                                                outputtw.Write("</" + thisTag + ">");
-                                            }
-                                            if (!isTagClose && currentTag.Length > 0)
-                                            {
-                                                if (!lengthReached)
-                                                {
-                                                    tagStack.Push(currentTag);
-                                                    outputtw.Write("<" + currentTag);
-                                                    if (!string.IsNullOrEmpty(tagContents))
-                                                    {
-                                                        if (tagContents.EndsWith("/"))
-                                                        {
-                                                            // No end tag e.g. <br />.
-                                                            tagStack.Pop();
-                                                        }
-
-                                                        outputtw.Write(tagContents);
-                                                        write = true;
-                                                        insideTagSpaceEncountered = false;
-                                                    }
-                                                    outputtw.Write(">");
-                                                }
-                                            }
-                                            // Continue to next iteration of the text reader.
-                                            continue;
-
-                                        default:
-                                            if (isInsideElement)
-                                            {
-                                                if (ic == (int)' ')
-                                                {
-                                                    if (!insideTagSpaceEncountered)
-                                                    {
-                                                        insideTagSpaceEncountered = true;
-                                                    }
-                                                }
-
-                                                if (!insideTagSpaceEncountered)
-                                                {
-                                                    currentTag += (char)ic;
-                                                }
-                                            }
-                                            break;
-                                    }
-
-                                    if (isInsideElement || insideTagSpaceEncountered)
-                                    {
-                                        write = false;
-                                        if (insideTagSpaceEncountered)
-                                        {
-                                            tagContents += (char)ic;
-                                        }
-                                    }
-
-                                    if (!isInsideElement || treatTagsAsContent)
-                                    {
-                                        currentTextLength++;
-                                    }
-
-                                    if (currentTextLength <= length || (lengthReached && isInsideElement))
-                                    {
-                                        if (write)
-                                        {
-                                            var charToWrite = (char)ic;
-                                            outputtw.Write(charToWrite);
-                                            currentLength++;
-                                        }
-                                    }
-
-                                    if (!lengthReached && currentTextLength >= length)
-                                    {
-                                        // Reached truncate limit.
-                                        if (addElipsis)
-                                        {
-                                            outputtw.Write("&hellip;");
-                                        }
-                                        lengthReached = true;
-                                    }
-
-                                }
-
-                            }
-                        }
-                    }
-                    outputtw.Flush();
-					outputms.Position = 0;
-					using (TextReader outputtr = new StreamReader(outputms))
-					{
-						return new HtmlString(outputtr.ReadToEnd().Replace("  ", " ").Trim());
-					}
-				}
-			}
+            return _stringUtilities.Truncate(html, length, addElipsis, treatTagsAsContent);
 		}
